Estructura del servicio confirmación reserva. Falta llamar a los servicios de los medios de pago, utilizar WS-Addressing y WS-Security

+++ TicketInco/src/main/java/com/fing/ticketinco/Reserva.java
@@ -12,7 +12,7 @@ import java.util.ArrayList;
  * @author javier
  */
 public class Reserva {
-    public int idReserva;    
+    public long idReserva;    
     public ArrayList<Horario> horarios;
     public int Estado;
 }
